move metadata builder to metadata folder
