+++ src/JMS/Serializer/SerializerBuilder.php
@@ -21,8 +21,8 @@ namespace JMS\Serializer;
 use Doctrine\Common\Annotations\AnnotationReader;
 use Doctrine\Common\Annotations\FileCacheReader;
 use Doctrine\Common\Annotations\Reader;
-use JMS\Serializer\Builder\DefaultDriverFactory;
-use JMS\Serializer\Builder\DriverFactoryInterface;
+use JMS\Serializer\Metadata\Builder\DefaultDriverFactory;
+use JMS\Serializer\Metadata\Builder\DriverFactoryInterface;
 use JMS\Serializer\Construction\ObjectConstructorInterface;
 use JMS\Serializer\Construction\UnserializeObjectConstructor;
 use JMS\Serializer\EventDispatcher\EventDispatcher;
