+++ tests/JMS/Serializer/Tests/Serializer/Doctrine/IntegrationTest.php
@@ -13,8 +13,8 @@ use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\Mapping\Driver\AnnotationDriver;
 use Doctrine\ORM\ORMException;
 use Doctrine\ORM\Tools\SchemaTool;
-use JMS\Serializer\Builder\CallbackDriverFactory;
-use JMS\Serializer\Builder\DefaultDriverFactory;
+use JMS\Serializer\Metadata\Builder\CallbackDriverFactory;
+use JMS\Serializer\Metadata\Builder\DefaultDriverFactory;
 use JMS\Serializer\Metadata\Driver\DoctrineTypeDriver;
 use JMS\Serializer\Serializer;
 use JMS\Serializer\SerializerBuilder;
